Changed commit command
I changed the command in step 4 to use -m to make commit messages.
This is because if you don't use -m and you are new, you will get
stuck in the nightmare of figuring out how to use less or vim on the
command line.  Fixing that problem will add 20 steps to this tutorial,
so it is easier for now to teach people the bad practice of using -m
than it is to teach them how to set the git default editor.